Extract implementation of Navigation to NavigationStack
Navigation contained all implementation details for navigating
between contexts

Extracting these implementation details to NavigationStack will
better adhere to the Dependency Inversion principle and make the
Navigation API more accessible to new developers
