+++ src/main/java/seedu/address/model/ModelManager.java
@@ -50,7 +50,7 @@ public class ModelManager implements Model {
 
         this.tracker = new Tracker(tracker);
         this.userPrefs = new UserPrefs(userPrefs);
-        this.navigation = new Navigation();
+        this.navigation = new NavigationStack();
         filteredModules = new FilteredList<>(this.tracker.getModuleList());
         lastListLevel = DisplayListLevel.MODULE;
     }
